fix(costs): accurate LLM cost via litellm price map + fractional cents
The Costs dashboard overstated DeepSeek spend ~50x. Two bugs:
- runtime.py used a hardcoded pricing table with wrong units (per-1M
  values applied with a per-1K divisor) while litellm's accurate
  completion_cost was computed and discarded.
- cost was stored as integer cents, truncating every sub-cent (i.e.
  most) real cost to zero.

New app/services/agent/cost.py resolves per-token rates from litellm's
model map, with a corrected fallback for models litellm doesn't ship
(deepseek-v4-flash). Cost columns widened to numeric(14,6) (migration
0005). Frontend shows sub-cent precision when spend is tiny.

+++ backend/app/api/knowledge.py
@@ -226,8 +226,8 @@ async def get_costs_summary(db: AsyncSession = Depends(get_db)):
 
     return {
         "total_tokens": int(total_tokens),
-        "total_cost_cents": int(total_cost),
-        "by_provider": [{"provider": p, "cost_cents": int(c)} for p, c in provider_result.all()],
+        "total_cost_cents": float(total_cost),
+        "by_provider": [{"provider": p, "cost_cents": float(c)} for p, c in provider_result.all()],
     }
 
 
@@ -242,7 +242,7 @@ async def get_costs_log(db: AsyncSession = Depends(get_db), limit: int = 20):
     logs = result.scalars().all()
     return [{"provider": l.provider, "model": l.model, "call_type": l.call_type,
              "tokens_in": l.tokens_in, "tokens_out": l.tokens_out,
-             "cost_cents": l.cost_cents, "date": str(l.date)} for l in logs]
+             "cost_cents": float(l.cost_cents), "date": str(l.date)} for l in logs]
 
 
 @router.get("/audit")
