+++ backend/app/models/audit.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String, Text
 from sqlalchemy import JSON
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -17,7 +17,9 @@ class AuditLog(Base, UUIDMixin):
     details: Mapped[dict | None] = mapped_column(JSON)
     llm_calls: Mapped[int | None] = mapped_column(Integer)
     tokens_used: Mapped[int | None] = mapped_column(Integer)
-    estimated_cost_cents: Mapped[int | None] = mapped_column(Integer)
+    # Fractional cents: real per-call costs for cheap models are sub-cent and
+    # must not truncate to zero.
+    estimated_cost_cents: Mapped[float | None] = mapped_column(Numeric(14, 6))
     started_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     completed_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     status: Mapped[str | None] = mapped_column(String(20))
@@ -41,4 +43,5 @@ class CostLog(Base, UUIDMixin):
     call_type: Mapped[str] = mapped_column(String(50))
     tokens_in: Mapped[int] = mapped_column(Integer, default=0)
     tokens_out: Mapped[int] = mapped_column(Integer, default=0)
-    cost_cents: Mapped[int] = mapped_column(Integer, default=0)
+    # Fractional cents (see AuditLog.estimated_cost_cents).
+    cost_cents: Mapped[float] = mapped_column(Numeric(14, 6), default=0)
